// jungsi-frontend/src/stores/server/features/explore/early-subject/interfaces.ts
export interface IExploreEarlySubjectStep3Item {
  id: number;
  university: {
    id: number;
    name: string;
    region: string;
    code: string;
    establishment_type: string;
  };
  admission: {
    id: number;
    name: string;
    year: number;
    basic_type: "일반" | "특별";
  };
  general_field: {
    id: number;
    name: string;
  };
  name: string;
  method: {
    method_description: string;
    subject_ratio: number | null;
    document_ratio: number | null;
    interview_ratio: number | null;
    practical_ratio: number | null;
    other_details: string | null;
    second_stage_first_ratio: number | null;
    second_stage_interview_ratio: number | null;
    second_stage_other_ratio: number | null;
    second_stage_other_details: string | null;
    eligibility: string;
    school_record_evaluation_score: string | null;
    school_record_evaluation_elements: string | null;
  };
}